requirements.txt
asgiref==3.9.1
certifi==2025.7.14
charset-normalizer==3.4.2
coreapi==2.3.3
coreschema==0.0.4
Django==5.2.4
django-cors-headers==4.7.0
djangorestframework==3.16.0
idna==3.10
itypes==1.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pillow==11.3.0
python-dotenv==1.1.1
requests==2.32.4
setuptools==80.9.0
sqlparse==0.5.3
uritemplate==4.2.0
urllib3==2.5.0
